+++ NeuralNetworkToolkit/ActivationFunctions/NNTKSigmoidActivationFunction.m
@@ -22,7 +22,7 @@
     return output * (1.0f - output);
 }
 
-- (void)compute:(float*)inputBuffer length:(int)length {
+- (void)compute:(float*)inputBuffer length:(NSUInteger)length {
     for (int i = 0; i < length; i++) {
         inputBuffer[i] = [self compute:inputBuffer[i]];
     }
